+++ src/components/Banner/Banner.jsx
@@ -1,11 +1,13 @@
 import $Banner from './Banner.styles.jsx';
 import Logo from './Logo/Logo.jsx';
+import Navigation from './Navigation/Navigation.jsx';
 
 export default function Banner () {
 
     return (
         <$Banner>
             <Logo />
+            <Navigation />
         </$Banner>
     )
 }