+++ src/components/Banner/Navigation/NavigationItem/NavigationItem.jsx
@@ -1,0 +1,11 @@
+import { $Icon, $NavigationItem } from './NavigationItem.styles.jsx';
+
+export default function NavigationItem ({ children, title }) {
+
+    return (
+        <$NavigationItem>
+            <$Icon>{children}</$Icon>
+            {title}
+        </$NavigationItem>
+    )
+}