+++ src/components/Main/Contact/Contact.jsx
@@ -1,0 +1,6 @@
+import { $Contact } from "./Contact.styles.jsx";
+
+export default function Contact() {
+
+    return (<$Contact>contact</$Contact>);
+}