+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.12 (2025-01-07)
+===============
+* [@BartoszDziurowicz](https://github.com/AthenaHouse/athena-web/pull/2)
+* Create and implement navigation menu with simple templates of child components
+* Align navigation with react router
+
 0.11 (2025-01-06)
 ===============
 * [@BartoszDziurowicz](https://github.com/AthenaHouse/athena-web/pull/1)